+++ tests/ImageSharp.Benchmarks/Bulk/ToVector4_Rgba32.cs
@@ -54,13 +54,13 @@ public class ToVector4_Rgba32 : ToVector4<Rgba32>
         Span<byte> sBytes = MemoryMarshal.Cast<Rgba32, byte>(this.source.GetSpan());
         Span<float> dFloats = MemoryMarshal.Cast<Vector4, float>(this.destination.GetSpan());
 
-        int n = dFloats.Length / Vector<byte>.Count;
+        nuint n = (uint)dFloats.Length / (uint)Vector<byte>.Count;
 
         ref Vector<byte> sourceBase = ref Unsafe.As<byte, Vector<byte>>(ref MemoryMarshal.GetReference((ReadOnlySpan<byte>)sBytes));
         ref Vector<float> destBase = ref Unsafe.As<float, Vector<float>>(ref MemoryMarshal.GetReference(dFloats));
         ref Vector<uint> destBaseU = ref Unsafe.As<Vector<float>, Vector<uint>>(ref destBase);
 
-        for (int i = 0; i < n; i++)
+        for (nuint i = 0; i < n; i++)
         {
             Vector<byte> b = Unsafe.Add(ref sourceBase, i);
 
@@ -75,10 +75,10 @@ public class ToVector4_Rgba32 : ToVector4<Rgba32>
             Unsafe.Add(ref d, 3) = w3;
         }
 
-        n = dFloats.Length / Vector<float>.Count;
+        n = (uint)(dFloats.Length / Vector<float>.Count);
         var scale = new Vector<float>(1f / 255f);
 
-        for (int i = 0; i < n; i++)
+        for (nuint i = 0; i < n; i++)
         {
             ref Vector<float> dRef = ref Unsafe.Add(ref destBase, i);
 
@@ -96,13 +96,13 @@ public class ToVector4_Rgba32 : ToVector4<Rgba32>
         Span<byte> sBytes = MemoryMarshal.Cast<Rgba32, byte>(this.source.GetSpan());
         Span<float> dFloats = MemoryMarshal.Cast<Vector4, float>(this.destination.GetSpan());
 
-        int n = dFloats.Length / Vector<byte>.Count;
+        nuint n = (uint)dFloats.Length / (uint)Vector<byte>.Count;
 
         ref Vector<byte> sourceBase = ref Unsafe.As<byte, Vector<byte>>(ref MemoryMarshal.GetReference((ReadOnlySpan<byte>)sBytes));
         ref Vector<float> destBase = ref Unsafe.As<float, Vector<float>>(ref MemoryMarshal.GetReference(dFloats));
         var scale = new Vector<float>(1f / 255f);
 
-        for (int i = 0; i < n; i++)
+        for (nuint i = 0; i < n; i++)
         {
             Vector<byte> b = Unsafe.Add(ref sourceBase, i);
